v1/requirements.txt
langchain
langchain-community
streamlit
sentence-transformers
faiss-cpu
openai
tiktoken
pyyaml
matplotlib
pandas
rouge
scikit-learn
pypdf
unstructured
python-docx
openpyxl
python-dotenv
PyPDF2 
python-docx 
pandas